+++ jukebox-app/build.gradle
@@ -1,6 +1,8 @@
-apply plugin: 'com.android.application'
-apply plugin: 'com.github.triplet.play'
-apply plugin: 'idea'
+plugins {
+    id 'com.android.application'
+    id 'com.github.triplet.play'
+    id 'idea'
+}
 
 def appVersionName = version ?: '1.0'
 def appVersionCode = Integer.valueOf(System.env.TRAVIS_BUILD_NUMBER ?: project.findProperty('versionCode') ?: 1)
