+++ jukebox-app/src/main/java/se/qxx/android/jukebox/activities/fragments/RemotePlayerFragment.java
@@ -1,41 +1,30 @@
 package se.qxx.android.jukebox.activities.fragments;
 
-import android.annotation.SuppressLint;
 import android.content.Intent;
-import android.media.MediaPlayer;
 import android.net.Uri;
 import android.os.Bundle;
-import android.os.Handler;
 import android.util.Log;
 import android.view.LayoutInflater;
-import android.view.SurfaceView;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.Button;
 import android.widget.ImageButton;
 import android.widget.LinearLayout;
-import android.widget.MediaController;
 import android.widget.ProgressBar;
 import android.widget.SeekBar;
 import android.widget.TextView;
 import android.widget.Toast;
 
-import com.google.android.gms.cast.MediaTrack;
 import com.google.android.gms.cast.framework.CastContext;
 import com.google.android.gms.cast.framework.Session;
 import com.google.android.gms.cast.framework.SessionManager;
 import com.google.android.gms.cast.framework.SessionManagerListener;
 import com.google.android.gms.cast.framework.media.RemoteMediaClient;
-import com.google.protobuf.RpcCallback;
 
 import org.apache.commons.lang3.StringUtils;
-
-import java.util.List;
 
 import se.qxx.android.jukebox.R;
 import se.qxx.android.jukebox.activities.SubSelectActivity;
 import se.qxx.android.jukebox.cast.ChromeCastConfiguration;
-import se.qxx.android.jukebox.cast.ChromeCastProvider;
 import se.qxx.android.jukebox.media.VideoControllerView;
 import se.qxx.android.jukebox.settings.JukeboxSettings;
 import se.qxx.android.jukebox.widgets.Seeker;
@@ -301,7 +290,7 @@ public abstract class RemotePlayerFragment extends PlayerFragment
         // stop the current cast provider
         if (this.isPlaying()) {
             this.stop();
-            initializeCastProvider(getView(), null, this, null);
+            //initializeCastProvider(getView(), null, this, null);
         }
 
         // start movie and seek?
